Refactoring & Bug Fixes.
